[FIX] mail: fix failing chat window test
This PR fixes the `chat window should open when receiving a new DM` test
that was sometimes failing. This test sends a message on the bus to
emulate the reception of a message. If this message is received before
messaging is initialized, the test fails. This PR fixes this issue by
waiting for messaging to be initialized.

fixes runbot-61162

closes odoo/odoo#160465

Signed-off-by: Didier Debondt (did) <did@odoo.com>